feat(windows): add memory metrics for windows
