+++ src/nri/sampler_all.go
@@ -1,0 +1,39 @@
+// Copyright 2025 New Relic Corporation. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build !windows
+
+package nri
+
+import "github.com/newrelic/nri-docker/src/biz"
+
+func memory(mem *biz.Memory) []entry {
+	metrics := []entry{
+		metricMemoryCacheBytes(mem.CacheUsageBytes),
+		metricMemoryUsageBytes(mem.UsageBytes),
+		metricMemoryResidentSizeBytes(mem.RSSUsageBytes),
+		metricMemoryKernelUsageBytes(mem.KernelUsageBytes),
+	}
+	if mem.MemLimitBytes > 0 {
+		metrics = append(metrics,
+			metricMemorySizeLimitBytes(mem.MemLimitBytes),
+			metricMemoryUsageLimitPercent(mem.UsagePercent),
+		)
+	}
+	if mem.SoftLimitBytes > 0 {
+		metrics = append(metrics, metricMemorySoftLimitBytes(mem.SoftLimitBytes))
+	}
+	if mem.SwapLimitBytes > 0 {
+		metrics = append(metrics, metricMemorySwapLimitBytes(mem.SwapLimitBytes))
+	}
+	if mem.SwapLimitBytes > 0 && mem.SwapLimitUsagePercent != nil {
+		metrics = append(metrics, metricMemorySwapLimitUsagePercent(*mem.SwapLimitUsagePercent))
+	}
+	if mem.SwapUsageBytes != nil {
+		metrics = append(metrics, metricMemorySwapUsageBytes(*mem.SwapUsageBytes))
+	}
+	if mem.SwapOnlyUsageBytes != nil {
+		metrics = append(metrics, metricMemorySwapOnlyUsageBytes(*mem.SwapOnlyUsageBytes))
+	}
+	return metrics
+}
